+++ capture/devconfig.py
@@ -1,19 +1,26 @@
 # devconfig.py
 
-#version control
-RoboVersion = 2.4
+import platform
+
+#######################################
+
+#  version control
+RoboVersion = 2.5
 
 # Hard coded limits
 max_robot_reagents = 7
 maxreagentchemicals = 3
-volspacing = 10 #reagent microliter (uL) spacing between points in the stateset 
+volspacing = 100 # reagent microliter (uL) spacing between points in the stateset
 
+#######################################
 
 # perovskite solvent list (simple specification of what is a liquid)
 # assumes only 1 liquid / reagent
-solventlist = ['GBL', 'DMSO', 'DMF', 'FAH', 'DCM'] #ya, I know FAH isn't a solvent, but it makes programming easier
+solventlist = ['GBL', 'DMSO', 'DMF', 'DCM']
 
 # lab file requirements list
+
+#######################################
 
 # Gdrive target folder for rendering
 targetfolder = '11vIE3oGU77y38VRSu-OQQw2aWaNfmOHe' #target folder for run generation
@@ -23,6 +30,18 @@ reagentsheetid = "1JgRKUH_ie87KAXsC-fRYEw_5SepjOgVt7njjQBETxEg"
 reagent_workbook_index = 1
 reagent_interface_amount_startrow = 15
 
+#######################################
+
+system = platform.system()
+
+if system == "Linux":
+    wolfram_kernel_path = '/usr/local/Wolfram/Mathematica/12.0/Executables/WolframKernel'
+# Mac
+elif system == "Darwin":
+    wolfram_kernel_path = None
+
+#######################################
+
 def labfiles(lab):
     if lab == "LBL" or lab == "HC":
         filereq = ['CrystalScoring','ExpDataEntry','metadata.json']
@@ -30,4 +49,9 @@ def labfiles(lab):
         filereq = ['CrystalScoring','metadata.json']
     return(filereq)
 
+######################################
+# Sampler
 
+# must be 'default' or 'wolfram'
+# 'wolfram' is currently experimental and unsupported
+sampler = 'default'