+++ capture/generate/qrandom.py
@@ -1,14 +1,18 @@
 import logging
 import optunity
 import pandas as pd
+import numpy as np
 import random
 import sys
 
+from capture.generate.wolframsampler import WolframSampler
 from capture.testing import inputvalidation
 from capture.models import chemical
 from capture.generate import calcs
+import capture.devconfig as config
 
 modlog = logging.getLogger('capture.generate.qrandom')
+
 
 def rdfbuilder(rvolmaxdf, rvolmindf, reagent, wellnum):
     count = 0 
@@ -17,33 +21,11 @@ def rdfbuilder(rvolmaxdf, rvolmindf, reagent, wellnum):
     while count < wellnum:
         volmax = rvolmaxdf[count]
         volmin = rvolmindf[count]
-#        _, info_random, _ = optunity.minimize(f, num_evals=1, x1=[volmin, volmax], solver_name='random search')
-#        reagentitem=info_random.call_log['args']['x1'] #Create reagent amounts (mmol), change x variable to change range, each generated from unique sobol index
         reagentitem = random.randint(volmin,volmax)
         reagentlist.append(reagentitem)
         count+=1
     rdf = pd.DataFrame({reagentname:reagentlist}).astype(int)
     return(rdf)
-
-## Determine the maximum concentration for a particular chemical in a reagent
-#def maxconcdet(rdata, rnum, currchemical):
-#    maxconc = 0
-#    for chemid,conc in rdata['%s' %rnum].concs.items():
-#        itemnum = chemid.split('m')[1]
-#        chemical_list_name = rdata['%s'%rnum].chemicals[itemnum]
-#        if currchemical in chemical_list_name:
-#            if conc > maxconc:
-#                maxconc = conc
-#    return(maxconc)
-#
-## Determine the absolute maximum concentration for a particular chemical across all reagents in a portion of the experiment
-#def absmaxcondet(rdata, rnum, currchemical, reagentlist):
-#    maxconc = 0 
-#    for reagent in reagentlist:
-#        conc = (maxconcdet(rdata, reagent, currchemical))
-#        if conc > maxconc:
-#            maxconc =  conc
-#    return(maxconc)
 
 def f(x1):
     return 0.0
@@ -65,42 +47,6 @@ def calcvollimit(userlimits, rdict, volmax, volmin, experiment, reagentlist, rea
     possiblemaxvolumes.append(volmax)
     possibleminvolumes = []
     possibleminvolumes.append(volmin)
-    # Determine if any user constraints override the maximum concentrations dependent on the volume limits
-#    for chemical in chemicals:
-#        #The code will update the maximum volume to meet the upper bound set by the user
-#        try: 
-#            maxconc = maxconcdet(rdict, reagent, chemical)
-#            absmaxconc = absmaxcondet(rdict, reagent, chemical, reagentlist)
-#            userdefinedmax = (userlimits['chem%s_molarmax'%chemical])
-#            # Update the constraints / limits on the total volume of this reagent to fit within the restrictions that the user might have set
-#            if userdefinedmax <= maxconc:
-#                volmaxnew = userdefinedmax / maxconc * volmax
-#                possiblemaxvolumes.append(int(volmaxnew))
-#            # If other reagents in the portion of the reaction can access higher concentrations, don't error out
-#            elif userdefinedmax <= absmaxconc:
-#               pass
-#            else: 
-#                #In the case where this upperbound cannot be met the user has overconstrained the system
-#                modlog.error("User defined concentration greater than physically possible for chemical%s in experiment %s" %(chemical, experiment))
-#                sys.exit()
-#        except Exception:
-#            pass
-#        #The code will update the minimum volume to meet the lower bound set by the user
-#        try: 
-#            reagconc = (rdata.concs['conc_chem%s' %chemical])
-#            userdefinedmin = (userlimits['chem%s_molarmin'%chemical])
-#            absmaxconc = absmaxcondet(rdict, reagent, chemical, reagentlist)
-#            if userdefinedmin >= 0:
-#                if reagconc < absmaxconc: 
-#                    pass
-#                else:
-#                    volminnew = userdefinedmin / reagconc * volmax
-#                    possibleminvolumes.append(int(volminnew))
-#            elif userdefinedmax <= absmaxconc:
-#                pass
-#            else: pass
-#        except Exception:
-#            pass
     # Take the values determined from the user constraints and pass along the relevant value to the calling function
     if (min(possiblemaxvolumes)) != volmax:
         # Flag that there are constraints so that the user is aware they are limiting the total physical space artificially
@@ -186,12 +132,90 @@ def calcvollimitdf(rdf, mmoldf, userlimits, rdict, volmax, volmin, experiment, r
     # Return the relevant datasets as int values (robot can't dispense anything smaller so lose the unsignificant figures /s)
     return(outvolmaxdf.astype(int), outvolmindf.astype(int))
 
-def portiondataframe(expoverview, rdict, vollimits, rxndict, wellnum, userlimits, experiment):
+
+def get_unique_chemical_names(reagents):
+    """Get the unique chemical species names in a list of reagents.
+
+    The concentrations of these species define the vector space in which we sample possible experiments
+
+    :param reagents: a list of perovskitereagent objects
+    :return: a list of the unique chemical names in all of the reagent
+    """
+    chemical_species = set()
+    for reagent in reagents:
+        chemical_species.update(reagent.chemicals)
+    return sorted(list(chemical_species))
+
+
+def build_reagent_vectors(portion_reagents, portion_chemicals):
+    """Write the reagents in a vector form, where the vectors live in the complete portion basis
+
+    The basis is spanned by all chemicals in all reagents in the portion (sans solvent)
+
+    :param portion_reagents: a list of all reagents in the portion
+    :param portion_chemicals: a list of all chemical names in the poriton
+    :param rdict:
+    :return:
+    """
+
+    # find the vector representation of the reagents in concentration space
+    reagent_vectors = {}
+    for reagent in portion_reagents:
+        name = 'Reagent{} (ul)'.format(reagent.name)
+        comp_dict = reagent.component_dict
+        vec = [comp_dict.get(elem, 0) for elem in portion_chemicals]
+        reagent_vectors[name] = vec
+
+    return reagent_vectors
+
+
+def wolfram_sampling(expoverview, rdict, vollimits, rxndict, wellnum, userlimits, experiment):
+    """Mike's implementation of a wrapper around Josh's Mathematica sampling code.
+    """
+    portionnum = 0
+    portion_mmol_df = pd.DataFrame()
+    experiment_mmol_df = pd.DataFrame()
+    experiment_df = pd.DataFrame()
+    ws = WolframSampler()
+    for portion in expoverview:
+
+        # Determine from the chemicals and the remaining volume the maximum and minimum volume possible for the sobol method
+        volmax = vollimits[portionnum][1]
+        # unoptimized code that ensure that the previous reagents are considered and that the final reagent accurately fills to the minimum volume set by the users "fill to" requirement
+
+        maxconc = rxndict.get('max_conc', 15)
+
+        portion_reagents = [rdict[str(i)] for i in portion]
+        portion_species_names = get_unique_chemical_names(portion_reagents)
+        reagent_vectors = build_reagent_vectors(portion_reagents, portion_species_names)
+        experiments = ws.randomlySample(reagent_vectors, int(wellnum), float(maxconc), float(volmax))
+        portion_df = pd.DataFrame.from_dict(experiments)
+        portion_df['Reagent6 (ul)'] = np.floor(portion_df['Reagent7 (ul)'] / 2)
+        portion_df['Reagent7 (ul)'] = np.ceil(portion_df['Reagent7 (ul)'] / 2)
+        rdict['6'] = rdict['7']
+        columnnames = portion_df.columns
+        for columnname in columnnames:
+            reagent = int(columnname.split('t')[1].split('(')[0])  # 'Reagent2 (ul)' to give '2'
+            mmol_df = calcs.mmolextension((portion_df[columnname]), rdict, experiment, reagent)
+            portion_mmol_df = pd.concat([portion_mmol_df, mmol_df], axis=1)
+
+        experiment_mmol_df = pd.concat([experiment_mmol_df, portion_mmol_df], axis=1)
+        experiment_df = pd.concat([experiment_df, portion_df], axis=1)
+        portionnum += 1
+
+    ws.terminate()
+    experiment_mmol_df.to_csv("mmol_df.csv")
+    experiment_df.to_csv("vol_df.csv")
+    return experiment_df, experiment_mmol_df
+
+def default_sampling(expoverview, rdict, vollimits, rxndict, wellnum, userlimits, experiment):
+    """Ian's original sampling implementation.
+    """
     portionnum = 0
     prdf = pd.DataFrame()
     prmmoldf = pd.DataFrame()
     for portion in expoverview:
-        # need the volume minimum and maximum and well count 
+        # need the volume minimum and maximum and well count
         reagentcount = 1
         reagenttotal = len(portion)
         # Determine from the chemicals and the remaining volume the maximum and minimum volume possible for the sobol method
@@ -222,7 +246,7 @@ def portiondataframe(expoverview, rdict, vollimits, rxndict, wellnum, userlimits
                 # The constraints on the middle draws are more complicated and are dependent upon the first, a different sampling strategy must be used (i.e. this is not going to use sobol as the ranges are different for each)
                 # Constrain the range based on volumne, reagent-chemical concentrations and user constraints
                 rvolmaxdf, rvolmindf = calcvollimitdf(finalrdf, mmoldf, userlimits, rdict, volmax, volmin, experiment, portion, reagent, wellnum, rxndict)
-                # Since each volume maximum is different, need to sample the remaining reagents independently (thus different sampling) 
+                # Since each volume maximum is different, need to sample the remaining reagents independently (thus different sampling)
                 rdf = rdfbuilder(rvolmaxdf, rvolmindf, reagent, wellnum)
                 mmoldf = calcs.mmolextension((rdf['Reagent%s (ul)' %reagent]), rdict, experiment, reagent)
                 reagentcount+=1
@@ -241,14 +265,14 @@ def portiondataframe(expoverview, rdict, vollimits, rxndict, wellnum, userlimits
                     rdf = rdfbuilder(rvolmaxdf, rvolmindf, reagent, wellnum)
                     mmoldf = calcs.mmolextension((rdf['Reagent%s (ul)' %reagent]), rdict, experiment, reagent)
                 reagentcount+=1
-            else: 
+            else:
                 modlog.error("Fatal error.  Unable to effectively parse reagent%s in portion %s.  Please make sure that the selected values make chemical sense!" %(reagent, portion))
             finalrdf = pd.concat([finalrdf, rdf], axis=1)
             finalmmoldf = pd.concat([finalmmoldf, mmoldf], axis=1)
         prdf = pd.concat([prdf, finalrdf], axis=1)
         prmmoldf = pd.concat([prmmoldf, finalmmoldf], axis=1)
-        portionnum+=1
-    return(prdf, prmmoldf)
+        portionnum += 1
+    return prdf, prmmoldf
 
 def ensuremin(rvolmindf, finalrdf, finalvolmin):
     currvoldf = finalrdf.sum(axis=1)
@@ -256,7 +280,7 @@ def ensuremin(rvolmindf, finalrdf, finalvolmin):
     rvolmindf[rvolmindf < trumindf] = trumindf
     return(rvolmindf)
 
-def preprocess(chemdf, rxndict, edict, rdict, climits):
+def preprocess_and_sample(chemdf, rxndict, edict, rdict, climits):
     ''' generates a set of random reactions within given reagent and user constraints
 
     requires the chemical dataframe, rxndict (with user inputs), experiment dictionary, 
@@ -278,8 +302,17 @@ def preprocess(chemdf, rxndict, edict, rdict, climits):
                     vollimits=(v)
                 else:
                     pass
-        modlog.info('Building reagent constraints for experiment %s using reagents %s for a total of %s wells' %(experiment, edict[experimentname], wellnum) )
-        prdf,prmmoldf = portiondataframe(edict[experimentname], rdict, vollimits, rxndict, wellnum, climits, experiment)
+
+        modlog.info('Building reagent constraints for experiment %s using reagents %s for a total of %s wells'
+                    % (experiment, edict[experimentname], wellnum))
+        if config.sampler == 'wolfram':
+            prdf, prmmoldf = wolfram_sampling(edict[experimentname], rdict, vollimits, rxndict, wellnum, climits, experiment)
+        elif config.sampler == 'default':
+            prdf, prmmoldf = default_sampling(edict[experimentname], rdict, vollimits, rxndict, wellnum, climits, experiment)
+        else:
+            modlog.error('Encountered unexpected sampler in devconfig: {}. Exiting.'.format(config.sampler))
+            sys.exit(1)
+
         erdf = pd.concat([erdf, prdf], axis=0, ignore_index=True, sort=True)
         ermmoldf = pd.concat([ermmoldf, prmmoldf], axis=0, ignore_index=True, sort=True)
         # Return the reagent data frame with the volumes for that particular portion of the plate